Fix issue with missing package.json
Unignore package.json and fix incorrect renaming logic.
Clean before building.

+++ Makefile
@@ -1,7 +1,7 @@
 rollup:
 	node ./_template/build-pkg.js rollup
 	# Restore Rollup config if it's hidden
-	if test -f _rollup.config.js; then _mv rollup.config.js _rollup.config.js; fi
+	if test -f _rollup.config.js; then mv _rollup.config.js rollup.config.js; fi
 	yarn
 
 webpack:
@@ -13,7 +13,8 @@ webpack:
 clean:
 	# Restore Rollup config if it's hidden
 	if test -f _rollup.config.js; then mv _rollup.config.js rollup.config.js; fi
-	rm package.json
-	rm yarn.lock
+	if test -f package.json; then rm package.json; fi
+	if test -f package-lock.json; then rm package-locki.json; fi
+	if test -f yarn.lock; then rm yarn.lock; fi
 
 .PHONY: rollup webpack clean
